refactor: move CustomId to core
