+++ packages/core/package.json
@@ -61,7 +61,8 @@
     "devDependencies": {
         "@seedcord/eslint-config": "workspace:*",
         "@seedcord/tsconfig": "workspace:*",
-        "@seedcord/tsdown-config": "workspace:*"
+        "@seedcord/tsdown-config": "workspace:*",
+        "type-fest": "catalog:deps"
     },
     "publishConfig": {
         "access": "public",
