+++ packages/core/src/customId/CustomId.ts
@@ -5,7 +5,7 @@ import { computeLayoutHash, decodeBody, encodeBody, HASH_LENGTH } from './codec'
 import { InvalidCustomId, StaleCustomId } from './Errors';
 
 import type { CustomIdField, CustomIdShape, DecodedParams } from './Field';
-import type { Snowflake } from 'discord.js';
+import type { Snowflake } from 'discord-api-types/v10';
 import type { NonEmptyTuple } from 'type-fest';
 
 // discord caps a customId at 100 chars.
