+++ packages/core/src/customId/Errors.ts
@@ -1,5 +1,5 @@
-import { Notice } from '@seedcord/core';
-import { NoticeCard } from '@seedcord/core/internal';
+import { Notice } from '@stops/Notice';
+import { NoticeCard } from '@stops/NoticeCard';
 
 import type { ReplyResponse } from '@seedcord/types';
