+++ packages/core/src/customId/codec.ts
