+++ packages/core/src/internal.index.ts
@@ -1,3 +1,7 @@
 export { setBotColor } from '@components/botColorHolder';
 
 export { NoticeCard } from '@stops/NoticeCard';
+
+export { prefixOf, decodeFor, type AnyCustomId } from '@customId/CustomId';
+export { ComponentDefsKey, type HasComponentDefs } from '@customId/routing';
+export type { CustomIdField, DecodedParams } from '@customId/Field';
