+++ packages/kit/src/pagination/cursor.ts
@@ -1,6 +1,6 @@
-import { CustomId } from '@customId/CustomId';
+import { CustomId } from '@seedcord/core';
 
-import type { CustomIdField } from '@customId/Field';
+import type { CustomIdField } from '@seedcord/core/internal';
 
 // Included in the layout hash so it is fixed per prefix. Reducing it marks every active button as StaleCustomId.
 // @internal
