+++ packages/seedcord/src/handlers/interaction/components/ButtonHandler.ts
@@ -1,6 +1,6 @@
 import { ComponentHandler } from './ComponentHandler';
 
-import type { AnyCustomId } from '@seedcord/kit/internal';
+import type { AnyCustomId } from '@seedcord/core/internal';
 import type { ButtonInteraction, CacheType } from 'discord.js';
 
 /**
