+++ packages/seedcord/src/miscellaneous/extractErrorResponse.ts
@@ -1,7 +1,7 @@
 import * as crypto from 'node:crypto';
 
 import { Notice, Fault } from '@seedcord/core';
-import { prefixOf } from '@seedcord/kit/internal';
+import { prefixOf } from '@seedcord/core/internal';
 import { Logger } from '@seedcord/services';
 import { DiscordAPIError } from 'discord.js';
