+++ packages/seedcord/tests/miscellaneous/extractErrorResponse.test.ts
@@ -1,5 +1,4 @@
-import { Fault } from '@seedcord/core';
-import { CustomId } from '@seedcord/kit';
+import { Fault, CustomId } from '@seedcord/core';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { extractErrorResponse, faultThrottle } from '@miscellaneous/extractErrorResponse';
