+++ packages/types/src/Interfaces/CustomId.ts
@@ -1,7 +1,7 @@
 /**
- * Structural shape of a customId matcher. A framework `CustomId` (from `@seedcord/kit`) satisfies it.
+ * Structural shape of a customId matcher. A framework `CustomId` (from `@seedcord/core`) satisfies it.
  * Used to type {@link InteractionsConfig.ignoreCustomIds} without importing the concrete class, which
- * would create a `@seedcord/types` to `@seedcord/kit` cycle.
+ * would create a `@seedcord/types` to `@seedcord/core` cycle.
  */
 export interface CustomIdMatcher {
     /** True if the raw customId wire was minted from this matcher's prefix. */
